% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{compute_outcomes}
\alias{compute_outcomes}
\title{Compute outcomes from modelOutput}
\usage{
compute_outcomes(modelOutput, severity_risk, ICU_risk, ventil_risks)
}
\arguments{
\item{modelOutput}{output returned by a model run}

\item{severity_risk}{vector of dimension 17, corresponding to the risk (proportion) of being a severe case, for each age group}

\item{ICU_risk}{vector of dimension 17, corresponding to the risk (proportion) of being a severe case, for each age group}
}
\description{
Compute outcomes from modelOutput
}
