% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{engine_run}
\alias{engine_run}
\title{Function that initialize (or reinitialize) the initial state of the model}
\usage{
engine_run(params, init)
}
\arguments{
\item{params}{list of various model parameters}

\item{init}{initial state of the model}
}
\value{
a matrix of the result
}
\description{
Function that initialize (or reinitialize) the initial state of the model
}
