% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HospiCoV-package.R
\docType{package}
\name{hospicov-package}
\alias{hospicov}
\alias{hospicov-package}
\title{hospicov: HospiCoV}
\description{
COVID-19 metapopulation model.
}
\details{
This package contains the whole HospiCoV model developped by EHESP.
Please refer to the vignettes to see how to use the package.
}
\author{
\strong{Maintainer}: Pascal Crépey \email{pascal.crepey@ehesp.fr}

Authors:
\itemize{
  \item Clément Massonnaud \email{clement.massonnaud@gmail.com}
  \item Jonathan Roux \email{jonathan.roux@ehesp.fr}
}

}
\keyword{internal}
