% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_inputs.R
\name{mod_inputs_ui}
\alias{mod_inputs_ui}
\alias{mod_inputs_server}
\title{mod_inputs_ui and mod_inputs_server}
\usage{
mod_inputs_ui(id)

mod_inputs_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
