% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map.R, R/mod_pop.R
\name{mod_map_ui}
\alias{mod_map_ui}
\alias{mod_map_server}
\alias{mod_pop_ui}
\title{mod_map_ui and mod_map_server}
\usage{
mod_map_ui(id)

mod_map_server(input, output, session)

mod_pop_ui(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.

A shiny Module.
}
\keyword{internal}
