% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model.R
\name{mod_model_ui}
\alias{mod_model_ui}
\alias{mod_model_server}
\title{mod_model_ui and mod_model_server}
\usage{
mod_model_ui(id)

mod_model_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
