% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{outcome_render}
\alias{outcome_render}
\title{Render stacked bar chart and table by age distribution}
\usage{
outcome_render(outcome_table, start_time, end_time, outcome = "severity")
}
\arguments{
\item{outcome_table}{the table returned by function compute_outcomes}
}
\description{
Render stacked bar chart and table by age distribution
}
