% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyHosp.R
\docType{class}
\name{PolyHosp}
\alias{PolyHosp}
\title{R6 Class representing a set of Parameters}
\format{\code{\link{R6Class}} object.}
\value{
Object of \code{\link{R6Class}} with all the population data.
}
\description{
R6 Class representing a set of Parameters

R6 Class representing a set of Parameters
}
\details{
Parameters contains all the parameters related to the population and the epidemic
}
\examples{
pop = PolyHosp$new()
}
\keyword{population}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nage}}{number of age groups}

\item{\code{vorHBase}}{hospital database containing all hospital information}

\item{\code{vorPolyHosp}}{the voronoi polygons}

\item{\code{vorPop}}{the population inside each polygons}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PolyHosp$new()}}
\item \href{#method-getMap}{\code{PolyHosp$getMap()}}
\item \href{#method-clone}{\code{PolyHosp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new `PolyHosp` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolyHosp$new(hospitals = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hospitals}}{list of hospitals on which we build the voronoi diagram}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PolyHosp` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMap"></a>}}
\subsection{Method \code{getMap()}}{
get the leaflet map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolyHosp$getMap()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a leaflet map with the polygons
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolyHosp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
