% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{runMod}
\alias{runMod}
\title{Wrapper function running the C++ model}
\usage{
runMod(params, sname, population, startDate = as.Date("2020-02-01"))
}
\arguments{
\item{params}{parameter object}

\item{sname}{name of the scenario}

\item{population}{population vector containing in the first cell the localization of the population}

\item{startDate}{the date of initial introduction of cases (Date object) (default: as.Date("2020-02-01"))}
}
\description{
Wrapper function running the C++ model
}
