% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{runMultiMod}
\alias{runMultiMod}
\title{Wrapper function running the C++ model multiple times}
\usage{
runMultiMod(paramsObj, sname, population, nbIterations)
}
\arguments{
\item{sname}{name of the scenario}

\item{population}{population vector containing in the first cell the localization of the population}

\item{params}{parameter object}
}
\description{
Wrapper function running the C++ model multiple times
}
