% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\docType{class}
\name{Parameters}
\alias{Parameters}
\title{R6 Class representing a set of Parameters}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} with all the parameters related to the population and the epidemic.
}
\description{
R6 Class representing a set of Parameters

R6 Class representing a set of Parameters
}
\details{
Parameters contains all the parameters related to the epidemic
}
\examples{
params = Parameters$new()
  
}
\keyword{parameter}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nage}}{number of age groups}

\item{\code{preImmune}}{the proportion of immune individuals at the beginning of the epidemic}

\item{\code{preExposed}}{the proportion of exposed individuals at the beginning of the epidemic}

\item{\code{preInfected}}{the proportion of exposed individuals at the beginning of the epidemic}

\item{\code{symptomatic}}{the proportion of symptomatic infection}

\item{\code{progression}}{the rate of transfer from E to I}

\item{\code{removal}}{the rate of transfer from I to R}

\item{\code{beta}}{the transmission probability upon contact with an infected (retro computed from R0)}

\item{\code{contact}}{the contact matrix}

\item{\code{agegroupnames}}{the name of the agegroups}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{susceptibility}}{A vector of size *nage* for susceptibility adjustment}

\item{\code{R0}}{the basic reproduction number}

\item{\code{duration}}{the duration fo the simulation i.e. c("Week", "Month", "Trimester", "Semester", "Year").}

\item{\code{nbDays}}{the number of days of the simulation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Parameters$new()}}
\item \href{#method-getList}{\code{Parameters$getList()}}
\item \href{#method-clone}{\code{Parameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `Parameters` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameters$new(R0 = 2.2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{R0}}{R0 (=3)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Parameters` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getList"></a>}}
\if{latex}{\out{\hypertarget{method-getList}{}}}
\subsection{Method \code{getList()}}{
Create a new list containing the required parameters for the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameters$getList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new list of parameters.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
