% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{runMod}
\alias{runMod}
\title{Wrapper function running the C++ model}
\usage{
runMod(params, sname, population)
}
\arguments{
\item{params}{parameter object}

\item{sname}{name of the scenario}

\item{population}{population vector}
}
\description{
Wrapper function running the C++ model
}
